LPS-83326 Move to the proper packages
