+++ portal-impl/src/com/liferay/portal/model/adapter/impl/StagedWorkflowDefinitionLinkImpl.java
@@ -17,8 +17,8 @@ package com.liferay.portal.model.adapter.impl;
 import com.liferay.exportimport.kernel.lar.StagedModelType;
 import com.liferay.portal.kernel.model.ClassName;
 import com.liferay.portal.kernel.model.WorkflowDefinitionLink;
+import com.liferay.portal.kernel.model.adapter.StagedWorkflowDefinitionLink;
 import com.liferay.portal.kernel.service.ClassNameLocalServiceUtil;
-import com.liferay.portal.kernel.workflow.adapter.StagedWorkflowDefinitionLink;
 import com.liferay.portal.model.impl.WorkflowDefinitionLinkImpl;
 
 import java.util.Date;
